[#1524] neofs-node: Reuse boltdb config for different components
