+++ cmd/neofs-node/config/engine/shard/metabase/config.go
@@ -1,20 +1,13 @@
 package metabaseconfig
 
 import (
-	"io/fs"
-
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-node/config"
+	boltdbconfig "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/engine/shard/boltdb"
 )
 
 // Config is a wrapper over the config section
 // which provides access to Metabase configurations.
 type Config config.Config
-
-// config defaults
-const (
-	// PermDefault is a default permission bits for metabase file.
-	PermDefault = 0660
-)
 
 // From wraps config section into Config.
 func From(c *config.Config) *Config {
@@ -37,18 +30,7 @@ func (x *Config) Path() string {
 	return p
 }
 
-// Perm returns the value of "perm" config parameter as a fs.FileMode.
-//
-// Returns PermDefault if the value is not a positive number.
-func (x *Config) Perm() fs.FileMode {
-	p := config.UintSafe(
-		(*config.Config)(x),
-		"perm",
-	)
-
-	if p == 0 {
-		p = PermDefault
-	}
-
-	return fs.FileMode(p)
+// BoltDB returns config instance for querying bolt db specific parameters.
+func (x *Config) BoltDB() *boltdbconfig.Config {
+	return (*boltdbconfig.Config)(x)
 }
