+++ cmd/neofs-node/config/engine/shard/writecache/config.go
@@ -2,6 +2,7 @@ package writecacheconfig
 
 import (
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-node/config"
+	boltdbconfig "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/engine/shard/boltdb"
 )
 
 // Config is a wrapper over the config section
@@ -132,4 +133,9 @@ func (x *Config) SizeLimit() uint64 {
 	}
 
 	return SizeLimitDefault
+}
+
+// BoltDB returns config instance for querying bolt db specific parameters.
+func (x *Config) BoltDB() *boltdbconfig.Config {
+	return (*boltdbconfig.Config)(x)
 }
